A base architecture for meta-modifying scripts done

+++ tirion.py
@@ -10,29 +10,34 @@ import signal
 
 file_name = './temp/tirion.job.%s.sh'%(int(time.time()))
 
-@atexit.register
-def cleanup():
-    print("Cleaning up...")
-    os.remove(file_name)    
-    exit(0)
-    
-#(iterations, !symmetry, types, output)
-signal.signal(signal.SIGINT, cleanup)
-signal.signal(signal.SIGTERM, cleanup)
-signal.signal(signal.SIGQUIT, cleanup)
-
 parser = argparse.ArgumentParser(prog='Tirion', description='A framework for STARCCM+ CFD Processing. Written for Formula Student Team Delft.')
 
 parser.add_argument('--name', metavar="-j", type=str, help='Job name.', required=True)
 parser.add_argument('--simulation', metavar="-s", type=str, help='Path to the .sim file to run processing on.', required=True)
-parser.add_argument('--iterations', metavar="-i", type=int, help='Number of iterations to run the simulation for.')
-parser.add_argument('--nodes', metavar='-n', type=int, help='Number of nodes to request the resource management system for.', required=True)
+parser.add_argument('--nodes', metavar='-n', default=1, type=int, help='Number of nodes to request the resource management system for.', required=True)
 parser.add_argument('--cores', metavar='-c', type=int, help='Number of cores per node.', required=True)
-parser.add_argument('--symmetry', type=bool, help='A flag to enable symmetry. [!stub(wip)]')
+parser.add_argument('--log', metavar='-l', default="", type=str, help="Path to which to save the logs from SLURM.")
+parser.add_argument('--temp', default=1, type=int, help="Should the generated temp job.sh file be saved")
+
+# WIP
+#(iterations, !symmetry, types, output) TODO: Merge mesh and pp scripts
+parser.add_argument('--iterations', metavar="-i", default=3000, type=int, help='Number of iterations to run the simulation for.')
 parser.add_argument('--output', metavar='-o', type=str, help='Ouptut directory')
+parser.add_argument('--symmetry', type=bool, help='A flag to enable symmetry. [!stub(wip)]')
 parser.add_argument('--type', metavar='-t', type=str, help='Type of processing: meshing | simulation | post-processing. Default is all.[!stub(wip)]')
 
 args = parser.parse_args()
+
+@atexit.register
+def cleanup():
+    print("Cleaning up...")
+    if args.temp == 1:
+        os.remove(file_name)
+    exit(0)
+
+signal.signal(signal.SIGINT, cleanup)
+signal.signal(signal.SIGTERM, cleanup)
+signal.signal(signal.SIGQUIT, cleanup)
 
 #The idea is to parse these and change .sh file, which you can open
 script = open('./tirion.sh','r')
@@ -42,20 +47,31 @@ script.close()
 # Change the arguments
 for i, line in enumerate(line_list):
     # Parse for the macro directory, job name, etc
-    if line[0:2] == "#&":
-        line_list[i] = "#&DEBEL\n"
-        
-    # Parse for macroPath, Simulation directory
-    if re.match("macroPath", line) is not None:
-        pass
+    if re.match("#STATUS", line) is not None:
+        line_list[i] = line.replace("[BASE]", "[META-MODIFIED]")
     
-    if re.match("simPath", line) is not None:
-        pass
+    if re.match("#SBATCH -J", line) is not None:
+        line_list[i] = line.replace("{?}" ,args.name)
+        
+    if re.match("#SBATCH --nodes", line) is not None:
+        line_list[i] = line.replace("{?}", str(args.nodes))
+        
+    if re.match("#SBATCH --n-tasks-per-node=", line) is not None:
+        line_list[i] = line.replace("{?}", str(args.cores))
+    
+    if re.match("#SBATCH -o", line) is not None:
+        line_list[i] = line.replace("{?}", args.log)
+
+    # Parse for macroPath, Simulation directory
+    if re.match("macroPath=", line) is not None:
+        macros = "./src/main/PostProcessing.java"
+        line_list[i] = line.replace("{?}", macros)
+    
+    if re.match("simPath=", line) is not None:
+        line_list[i] = line.replace("{?}", args.simulation)
 
 file_to_write = open(file_name, 'w')
 file_to_write.writelines(line_list)
 file_to_write.close()
 
-# _ = None
-# call slurm to submit the job and print the id, does slurm actually need the file, or is it copied?
-# subprocess.call("bash -c './tirion.sh %s %s %s %s %s'" % (args.simulation, _, _, _, _), shell=True)
+subprocess.call("sbatch %s" % (file_name), shell=True)
